Address NavFn review feedback

+++ nav2_navfn_planner/include/nav2_navfn_planner/parameter_handler.hpp
@@ -39,8 +39,8 @@ struct Parameters
   double tolerance;
   // Whether to use the astar planner or default dijkstras
   bool use_astar;
-  // Factor applied to the largest costmap dimension to limit path extraction cycles
-  double max_cycles_factor;
+  // Integer factor applied to the largest costmap dimension to limit path extraction cycles
+  int max_cycles_factor;
 };
 
 /**
